Update RSpec (revision 3014)
To coincide with the update to Rails 2.0RC2, update the copy of the RSpec
trunk that's currently installed inside the application.

Note that in the upgrade notes I've used "git rm -r" rather than "rm -r"
to blow away the old version as it makes adding the new files back in a
little easier.

Signed-off-by: Wincent Colaiuta <win@wincent.com>