New models, lib additions and refactoring
This is the "disappear into a cave" and come out days later with a bunch
of changes commit. I really would have liked to break this up into
chunks but the changes are far too intermixed to make that very
straighfoward.

So in this commit we have a host of new stuff: new CSS, more view
templates, the "acts as sortable" module, a refactoring of the
authentication module that makes it far more elegant, changes to the
links model and controller, additional graphics, notes, the
beginnings of a bayesian classifier, and some a "searchable" module
(think "acts as ferret").

From here on will try to be more disciplined and break up the changes a
little more!

Signed-off-by: Wincent Colaiuta <win@wincent.com>