Drop Revisions model
My initial wiki implementation only had a "revisions" notion because I
unthinkingly modelled it after other wikis. In reality I don't need it
and it is an unecessary complication, so remove it. (This is not a
community-driven wiki, nor one which allows anonymous edits, so the
ability to go back in time is not as important).

Signed-off-by: Wincent Colaiuta <win@wincent.com>