Set up navbar
In 2.0 spirit we're going for the simplest navigation possible. Some of
the CSS exhibits some mysterious behaviour (namely, I had to set a
-0.35em left margin on the links in the list items in order to get the
highlighting to correctly extend all the way to the preceding border; I
have no idea why this hack was required) but it displays correctly in
both Firefox and Safari.

Signed-off-by: Wincent Colaiuta <win@wincent.com>