Session/login enhancements
Rather than regenerating the session key on every single page access
regenerate it only on logging in (as before, destroy it on logging out).

By regenerating immediately on login we add another layer of
defense-in-depth against session fixation attacks (which are already
pretty unlikely due to the other layers), but we also reduce load on
the database (because the regenerated keys no longer need to be stored
on every page access) and we also avoid possible problems with
concurrent page accesses (if there are multiple pages open and
navigating in each of them tries to update the session key
independently).

Signed-off-by: Wincent Colaiuta <win@wincent.com>