Make wincent_on_rails the repository root
This is a huge reorganization which makes the wincent_on_rails app the
root of the repository and moves the items which were previously at the
top level into the "other" subdirectory.

The motivation for this change is that it makes deployment with
Capistrano easier, because Capistrano assumes that the application to
be deployed resides in the respository root (this is at least the case
with Git repositories; for Subversion repositories all you have to do
is tweak your repository URL to point to the desired subdirectory).

There are two alternatives to this change: patch Capistrano to teach it
to accept parameters specifying a subdirectory; or modify the deployment
script to override the built-in tasks provided by Capistrano and act
with subdirectories in mind. The former was rejected because I am on a
tight timeline, don't have time to learn the Capistrano codebase, and
don't know whether such a patch would be accepted anyway, and the latter
because it is brittle, error-prone and would also require a fairly
intimate knowledge of the inner workings of the default tasks proided by
Capistrano. So although the first alternative (enhancing Capistrano) is
the most elegant, I instead opted to restructure my repository (out of
pragmatism).

Signed-off-by: Wincent Colaiuta <win@wincent.com>