Add sample deploy files to test_app
This is the deploy app that I tried in my initial Capistrano testing
with the test app. It implements the basics, and things like
"cap deploy:prepare" and "cap deploy:check" already work, although
"cap deploy" does not.

Now that I've restructured the repository (in 581d1e1) I can't really
do test deployments with the test app any more, and have to do them
with the real app, so this file is now most of historical interest.

Signed-off-by: Wincent Colaiuta <win@wincent.com>