Capistrano: new tasks: migrate_all and cold override
The "migrate_all" tasks runs migrations for the test, development and
production databases and sets the environment up appropriately so that
the migrations run with the correct privileges (this is required in
the case of the production database because the normal Rails users does
not have adequate privileges to run migrations on it).

At the same time introduce a "cold" recipe override that calls this
"migrate_all" task instead of the default "migrate" one, which won't
work.

Signed-off-by: Wincent Colaiuta <win@wincent.com>