Upgrade to Haml 1.8.2, unpacking the gem into "vendor"
The Haml team want you to install Haml as a gem and then run
"haml --rails ." to stick a stub file in your "vendor/plugins"
directory. This unfortunately doesn't allow you to tie your
application to a known-good version of Haml.

I've gotten around this in the past by just doing an "svn export" of the
desired Haml version directly into "vendor/plugins", but now that the
Haml team has switched to Git that's no longer viable.

So now I've done a "gem unpack", sticking 1.8.2 in a new "vendor/gems"
directory and tweaking the environment (adding to the load path and
explicitly requiring Haml) so that things should continue to work.

Signed-off-by: Wincent Colaiuta <win@wincent.com>