Move Haml back into plugins folder
So looks like Haml can't be installed into vendor/gems without ugly
hacks. The problem is that if you modify the load_paths in the Rails
initializer then it will see Haml's init.rb file and try to run it,
but too soon (at least, that is what I believe is happening) and
this causes the server to fail to start.

The easiest solution is to just move the unpacked gem back into the
plugins directory and volia, it works.

Signed-off-by: Wincent Colaiuta <win@wincent.com>