Disallow slash in wiki article titles
The slash is a route delimiter, so explicitly disallow it in wiki
article titles. This bug was discovered while browsing the results of
the MediaWiki import. An article like "foo/bar" could be imported but
never edited because route recognition would fail for
"/wiki/foo%2Fbar/edit".

The problem is evidently that the %2F is getting converted to / _before_
route recognition takes place, thus breaking recognition. As this is out
of my hands (this is a Rails problem) I am going to disallow slashes in
article titles for now.

Signed-off-by: Wincent Colaiuta <win@wincent.com>