Wiki index improvements
The wiki index now shows "recently updated" articles instead of the
misleadingly named "recent articles", and shows a mini tag cloud with
the top 10 tags within the wiki. The "back" links have been renamed to
"index" because the former may be misleading depending on how you got
to the page containing the link in the first place.

Signed-off-by: Wincent Colaiuta <win@wincent.com>