Use FixtureReplacement to create new users in sessions story
This should be less brittle than calling ActiveRecord directly (in which
case we'd have to worry about valid attributes). By using
FixtureReplacement we can focus on only the details that are pertinent to
the scenario.

The down side is that the has_many/belongs_to relationship forces us to
do things "inside out" (ie. create_email comes before create_user even
though the _evaluation_ order is the opposite); this is unlike doing it
in Rails, where User.create comes first and the "build" method is used
to add the email afterwards. I am still not sure what to do about this,
and whether it is worth looking at patching FixtureReplacement to make
this pattern easier (does it occur often enough?).

Signed-off-by: Wincent Colaiuta <win@wincent.com>