More natural order for user creation in sessions story
Following on from 2e31d2d, I think this is a more "natural" order for
user creation, even though it does turn it into a two-line, two-step
process. This is a better match for the standard Rails pattern, which
is also a two-line, two-step process in which we first create the user
and only then add the email address.

Signed-off-by: Wincent Colaiuta <win@wincent.com>