Add "last_commenter" field to commentable models
This is principally required by for forum topics, but to make the
"commentable" interface consistent I have added it to all four
commentable classes.

Basically, whenever we create or destroy a comment we now update
a "last_commenter" cache in the associated "commentable" table.

Note that in defining an after_create filter we effectively break
the built-in counter cache (although we don't break the corresponding
before_destroy callback because we define after_destroy instead);
rather than fixing something that is half-broken, and which could
become fully-broken or not-broken-at-all in a future Rails update, we
instead turn off the built-in counter cache and roll our own. In any
case this is more efficient because it allows us to update both the
"last commenter" and the counter cache with a single database query.

In a future commit I will consider updating the "updated_at" field as
well; this is useful in the forums because you want topics to be
considered updated whenever a comment is added, but it is not so
useful in weblog posts, for example. So the most likely path I'll
take is adding another field, "last_commented_at" to all the commentable
models.

Signed-off-by: Wincent Colaiuta <win@wincent.com>