AJAX-ify comment destruction, confirm destructive operations
The "destroy" button was previously a normal form but now we do it via
AJAX. We preserve the confirmation dialog, and also add one to the
"spam" button.

I may later decide to remove the confirmation on the "spam" button
because it is not an irreversibly destructive operation like "destroy"
is (ie. the comment disappears from the page but it isn't actually
deleted). I guess it depends on the amount of spam I get whether the
confirmations will be a hinderance or a help.

Signed-off-by: Wincent Colaiuta <win@wincent.com>