Big reworking of comment callbacks
Add a very simple, optional callback mechanism for commentable classes
to indicate whether they want their timestamps updated whenever comments
are added or removed.

For some classes like (forum) topics and issues, it makes sense for them
to be considered "updated" whenever a comment is added or removed. In
this way we can simply get a list of most recent topics or issues with
a simple SELECT query that uses "ORDER BY updated_at".

In contrast, classes like (weblog) posts and (wiki) articles shouldn't
have their "updated" timestamp modified when comments are added or
removed. We want to be able to SELECT them and "ORDER BY updated_at" and
have their order unaffected by the comments on them.

Signed-off-by: Wincent Colaiuta <win@wincent.com>