Tackle N + 1 SELECT problem in forums controller
Here we replace the old query with a new one that avoids the N + 1
problem but I've actually left the older query, and an intermediate one,
commented out in the method for documentation purposes.

In effect the three queries are "N + 1", "N + some" and "N + none", and
they represent a trade off between lots of simple queries and fewer
complex ones. I'll need to benchmark this on the production server to
decide which one performs best (here on the heavily-loaded, slow local
machine in the development environment they are all equally slow).

Signed-off-by: Wincent Colaiuta <win@wincent.com>