Use destroy in Forum model
In 4988c63 I switched from using :delete_all to :destroy in both the
Forum and Topic models, but this was going a little too far. It turns
out that the most efficient pattern is to use :destroy at the Forum
level (so it can trigger the callbacks when the associated Topic
instances are destroyed), and use :delete_all at the Topic level (so
that it can blow away all the comments as quickly as possible).

As the Comment instances themselves have no dependent records blowing
them away with :delete_all doesn't leave any cruft behind in the
database.

Signed-off-by: Wincent Colaiuta <win@wincent.com>