Up date comments_count cache in User model
Use callbacks to update the comments_count counter cache in the User
model whenever a comment is created or destroyed.

At the same time set the Topic model to :destroy dependent comments
rather than :delete_all, because now that we have this cache we actually
want to ensure that it gets updated if we destroy a topic.

Signed-off-by: Wincent Colaiuta <win@wincent.com>