Add "heartbeat" controller
At the moment monit is hitting the front page every 30 seconds to make
sure everything is okay, but this is filling up the log file with a lot
of noise (4 hits, 1 for each mongrel, every 30 seconds, and each hit is
a slab of text 10 lines long.

So set up a lightweight "heartbeat" controller that does no logging but
which does exercise the entire stacking (routing, the controller itself,
sessions, a database query, a template render, and Haml).

At the same time we update the monit config to hit the new controller
and action, as well as check for some text that we expect to see on the
template.

Signed-off-by: Wincent Colaiuta <win@wincent.com>