Fix logging broken by heartbeat controller
Doing a "self.logger = nil" ending up silencing _all_ logging for the
app. This alternative seems to be the only way to silence logging for
only one controller.

It's still not ideal (I'd like to see error conditions still logged,
for instance) but it is the best compromise for now.

Signed-off-by: Wincent Colaiuta <win@wincent.com>