More efficient finds in topics controller
As always, as soon as you start nesting resources in any way you
have to deal with "N + 1 SELECT" issues. In this case we try to
make it as simple as possible by separately fetching the matching
comments rather than doing an ugly, heavyweight :include across
multiple levels of nesting.

Signed-off-by: Wincent Colaiuta <win@wincent.com>