Capitalize article titles when not found
This applies to admin users only: when an article is not found and we
offer to create it, capitalize the first letter as we almost certainly
are going to want the article title to start with a capital, yet often
the not-found links will be in lowercase.

For example, given input like "I like [[tagging]]", if the "tagging"
article does not exist and the admin clicks on the link, he/she will
most likely want the title to be "Tagging".

Signed-off-by: Wincent Colaiuta <win@wincent.com>