Experimental attempt at page caching wiki index (HTML)
This won't work in practice seeing as this is a paginated resource, so
it's commented out.

The only way to make this work would be to use a nested resource for
pages (eg. /wiki/12 or /wiki/pages/1 etc) but this would clash with
articles of the same name.

So I suspect that page-caching is a no-go for this action (and for
any paginated action, really) and so the only alternative will be
fragment caching.

Signed-off-by: Wincent Colaiuta <win@wincent.com>