Add RestfulPaginator class and clean up Paginator
RestfulPaginator is a subclass of the Paginator class design for
"RESTful" pagination (pagination where a "page" is modelled as a
collection on a non-singleton resource).

Given an articles model (aliased to "wiki") this gives us URLs like the
following:

  /wiki          # this index
  /wiki/foo      # the "foo" article
  /wiki/34       # page 34 of the index

Rather than:

  /wiki?page=34  # page 34 of the index

This comes at the cost of two things: one, you can never have articles
whose names are only numbers, and; two, you need to add an additional
map.connect to your routes file with a higher precedence than the
map.resources so that the page number URLs can be caught and routed to
the #index action rather than the #show action.

In my case these costs are acceptable (I don't anticipate ever needing
number-only article titles). As you can see, the articles model already
uses human-readable permalinks rather than internal database IDs, so
these pagination URLs are fine.

The whole motivation behind this addition is that it makes paginated
indexes page-cacheable, because now we'll have nice cached page paths
like "/wiki.html" (for the first page) and "/wiki/2.html" (for
subsequent pages).

For now I'm going to teach the wiki controller to use the new paginator,
and I may later do the same for the posts controller (index action).
This technique is suitable for use wherever the resource uses non-numeric
resource IDs (ie it will work for the post controller and probably the
forums controller as well), wherever page caching is a key goal, and
wherever you don't plan on offering sortable table headers (which you
most likely aren't going to do if you are page caching anyway).

Signed-off-by: Wincent Colaiuta <win@wincent.com>