Back out restful pagination and page caching for the wiki index
This commit largely reverts 0575803. I've had to back it out because I
just realized that the wiki index shows relative dates on it, which
means that it is not suitable for page caching. If performance ever
becomes a problem I can easily turn off relative dates and roll out
page caching again.

This may be a lesson in "premature optimization being the root of all
evil".

Signed-off-by: Wincent Colaiuta <win@wincent.com>