Scope wiki "top tags" to wiki articles only
Previously the wiki's "top tags" listing was imperfect because although
the tags really were the top tags for the wiki (ie. scoped to the wiki)
the links took the user to a globally-scoped tags page.

In other words, if "foo" was the top tag in the wiki, clicking on it
would take the user to a page showing all items tagged with "foo", not
just wiki articles. This may or may not be desirable behaviour, so as an
experiment I'm trying to scope the links too.

Signed-off-by: Wincent Colaiuta <win@wincent.com>