Add "all items tagged with" link to the show_subset template
If a user clicks on a link that shows all articles tagged with "foo",
provide a link which allows them to see all items (not just articles)
tagged with "foo".

I am still not convinced that this kind of scoping makes for the best UI
but this commit is a minor improvement anyway.

Signed-off-by: Wincent Colaiuta <win@wincent.com>