Tag search and linking user interface improvements
This is a nice resolution to the dilemma about what to do with
scoped tag searches and links. If a user clicks on one of the
tags listed under "Top tags in wiki", instead of being shown
only wiki articles, they will be shown all matching item but
with wiki articles at the top.

Signed-off-by: Wincent Colaiuta <win@wincent.com>