Email deletion UI and related improvements
Hook up the email deletion (and addition UI). The admin can now delete
and undelete emails, and uses can add and delete emails. Confirmation
emails are sent when a user adds an email.

I've had to drop the nested confirmation URLs because they stopped
working when I made pretty URLs for editing user emails (for the admin);
this was caused by the to_param changes in the Email model. This is
less secure, but a brute force attack is still improbable given the
size of the key space and the amount of time available before
confirmation keys expire. A side-effect is that confirmation URLs are
shorter now and less likely to be broken by clients which do extreme
wrapping.

Signed-off-by: Wincent Colaiuta <win@wincent.com>