Prettify group names in tag search results
Make the search results a little more user-friendly by converting
"article" into "wiki article" and "post" into "weblog post". Users
shouldn't need to know the meaning of the internal model names and
may be confused by them (what is the difference between an "article"
and a "post"?), especially in cases like Article and Post which are
routed using totally different names in the URLs ("wiki" and "blog"
respectively).

Signed-off-by: Wincent Colaiuta <win@wincent.com>