Passphrase resetting and mailer overhaul
The most important change here is the implementation of passphrase
resetting. Rather than mailing a new random passphrase (which a user
will likely forget) we instead mail a secret URL that can be used to
set a new passphrase. We guard against brute force attacks (URL
guessing) by requring the user to enter the email address as well.

Other changes include:

- pass request into reset mailer and confirmation mailer so that it
can dynamically generate the right target URL rather than hard-coding it
to wincent.com (which is broken becaus it makes it impractical to
deploy the same code to the staging and production environments)

- get rid of unnecessary optional "sent_at" parameter in mailer methods;
we also want "sent_at" to be "now"

- extend User model with new "resetting_passphrase" attribute; this
allows us to temporarily override the default validation behaviour (we
usually want to make sure the "old_passphrase" is correct for passphrase
updates but this is not the case when doing "resets" because the user
has presumably forgotten their old passphrase at that point)

- remove ignored return values from validates_each block in User model

- extract "find_by_email" class method in User model for better
readability and re-use by the resetting machinery

- clean up the User.authenticate method for better readability

- reword login template; "Forgot your passphrase?" sounds a little
friendlier than "Lost password"

Signed-off-by: Wincent Colaiuta <win@wincent.com>