Show forums with no topics in index
We need special handling of forums with no topics in them yet. The LEFT
OUTER JOIN ensures that a row is returned for all forums, not just those
with topics, and the modifications to the view and the helper ensure we
handle NULL/nil entries appropriately when we get forums without topics.

Signed-off-by: Wincent Colaiuta <win@wincent.com>